requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
jinja2
stripe
python-multipart
